Tighten GFQL temporal helper split
