+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - **GFQL policy / Cypher compiler hooks (#1454)**: Added experimental exact-key `precompile` and `postcompile` policy hooks for local Cypher string-query compilation. `postcompile` reports success or failure using the existing policy `success`, `error`, and `error_type` fields plus a stable `CompileSummary` with scalar compiler metadata.
 
 ### Changed
-- **GFQL temporal text helper split (#1498)**: Split temporal value parsing, duration parsing/folding, truncation folding, and AST rewrite helpers into `graphistry.compute.gfql.temporal` modules while keeping the existing `graphistry.compute.gfql.temporal_text` import surface stable. No temporal semantics changed.
+- **GFQL temporal text helper split (#1498)**: Split temporal constructor parsing, value parsing, duration parsing/folding, truncation folding, and AST rewrite helpers into `graphistry.compute.gfql.temporal` modules while keeping `graphistry.compute.gfql.temporal_text` as a small compatibility shim. No temporal semantics changed.
 - **GFQL / Cypher path-alias carry multiple MATCH native planning (#1495)**: shortestPath queries that carry an unused named path alias into a follow-on `MATCH` now compile with a logical plan and dispatch through the physical same-path route instead of the final `multiple_match_stages` residual chain fallback. Path-value references such as `length(path)` after a follow-on `MATCH` now fail fast rather than relying on unsupported residual-chain path materialization.
 - **GFQL / Cypher modularization audit cleanup (#1058)**: Removed an unused private aggregate-alias helper from Cypher lowering after auditing the large GFQL engine files for narrow safe deletion opportunities.
 - **GFQL / Cypher shortestPath endpoint-binding multiple MATCH native planning (#1489)**: shortestPath queries whose earlier `MATCH` clauses only bind endpoint nodes now receive a logical plan and enter native physical dispatch instead of the residual `multiple_match_stages` chain fallback. Path-alias-carrying follow-on `MATCH` shapes remain explicitly classified under `multiple_match_stages`.
