+++ graphistry/compute/gfql/row/ordering.py
@@ -18,7 +18,7 @@ from graphistry.compute.gfql.series_str_compat import (
     series_str_fullmatch,
     series_str_match,
 )
-from graphistry.compute.gfql.temporal_text import (
+from graphistry.compute.gfql.temporal.constructors import (
     DATETIME_CALL_TEXT_RE,
     DATE_CALL_TEXT_RE,
     LOCALDATETIME_CALL_TEXT_RE,
