+++ graphistry/compute/gfql/temporal/durations.py
@@ -7,7 +7,7 @@ from datetime import timezone as py_timezone
 import re
 from typing import Optional, cast
 
-from graphistry.compute.gfql import temporal_text as _tt
+from graphistry.compute.gfql.temporal import constructors as _tt
 from graphistry.compute.gfql.expr_parser import ExprNode, Literal
 from graphistry.compute.gfql.temporal.values import (
     _TemporalValue,
